fix exception for particle effect at the begining of network game

+++ src/units/effect.h
@@ -149,7 +149,7 @@ struct TargetParticleType
 	int Color;
 	Vector vR,vT,vD;
 	int s;
-	short pDist, LifeTime;
+	int pDist, LifeTime;
 	void aQuant(void);
 	void aQuant2(void);
